link-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
nanoid  # For generating short codes
python-jose[cryptography]